+++ medicalcompany-appointments/pom.xml
@@ -25,6 +25,11 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>com.personal</groupId>
+			<artifactId>medicalcompany</artifactId>
+			<version>1</version>
+      </dependency>
 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
