+++ medicalcompany-appointments/src/main/java/com/personal/medical/appointments/model/DentalAppointment.java
@@ -1,9 +1,13 @@
-package com.personal.medical.model;
+package com.personal.medical.appointments.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+import com.personal.medical.model.Dentist;
+import com.personal.medical.model.Patient;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -13,10 +17,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Entity
 @EqualsAndHashCode(callSuper=false)
-public class DoctorAppointment extends Appointment {
+public class DentalAppointment extends Appointment {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
+	
+	@ManyToOne
+	private Dentist dentist;
+	
+	@ManyToOne
+	private Patient patient;
 
 }
