+++ medicalcompany-appointments/src/main/java/com/personal/medical/appointments/model/DoctorAppointment.java
@@ -1,10 +1,9 @@
-package com.personal.medical.model;
+package com.personal.medical.appointments.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -14,16 +13,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Entity
 @EqualsAndHashCode(callSuper=false)
-public class DentalAppointment extends Appointment {
+public class DoctorAppointment extends Appointment {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	
-	@ManyToOne
-	private Dentist dentist;
-	
-	@ManyToOne
-	private Patient patient;
 
 }
