+++ medicalcompany-appointments/src/main/java/com/personal/medical/appointments/services/CrudService.java
@@ -1,0 +1,12 @@
+package com.personal.medical.appointments.services;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface CrudService<T, ID> {
+
+	T save(T t);
+	void delete(T t);
+	List<T> findAll();
+	Optional<T> findById(ID id);
+}
