+++ medicalcompany-appointments/src/main/java/com/personal/medical/appointments/services/DentalAppointmentService.java
@@ -1,0 +1,8 @@
+package com.personal.medical.appointments.services;
+
+import com.personal.medical.appointments.model.DentalAppointment;
+
+public interface DentalAppointmentService extends CrudService<DentalAppointment, Long> {
+	
+	
+}
