+++ medicalcompany-appointments/src/main/java/com/personal/medical/appointments/services/DoctorAppointmentService.java
@@ -1,0 +1,8 @@
+package com.personal.medical.appointments.services;
+
+import com.personal.medical.appointments.model.DoctorAppointment;
+
+public interface DoctorAppointmentService extends CrudService<DoctorAppointment, Long> {
+	
+	
+}
