+++ medicalcompany-appointments/src/test/java/com/personal/medical/appointments/services/DoctorAppointmentServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.personal.medical.services;
+package com.personal.medical.appointments.services;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.not;
@@ -13,23 +13,23 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.personal.medical.model.DentalAppointment;
-import com.personal.medical.repository.DentalAppointmentRepository;
+import com.personal.medical.appointments.model.DoctorAppointment;
+import com.personal.medical.appointments.repositories.DoctorAppointmentRepository;
 
 @SpringBootTest
-class DentalAppointmentServiceImplTest {
+class DoctorAppointmentServiceImplTest {
 
-	private DentalAppointmentService service;
+	private DoctorAppointmentService service;
 	
 	@Mock
-	private DentalAppointmentRepository repository;
+	private DoctorAppointmentRepository repository;
 	
 	@Mock
-	private DentalAppointment dentalAppointmentMock;
+	private DoctorAppointment doctorAppointmentMock;
 	
 	@BeforeEach
 	void setUp() throws Exception {
-		service = new DentalAppointmentServiceImpl(repository);
+		service = new DoctorAppointmentServiceImpl(repository);
 	}
 
 	@Test
@@ -40,9 +40,9 @@ class DentalAppointmentServiceImplTest {
 	@Test
 	void test_FindById_ReturnsMock_WhenCalledWithId1() {
 		//Arrange
-		when(repository.findById(1L)).thenReturn(Optional.of(dentalAppointmentMock));
+		when(repository.findById(1L)).thenReturn(Optional.of(doctorAppointmentMock));
 		//Assert
-		Assertions.assertEquals(dentalAppointmentMock, service.findById(1L).get());
+		Assertions.assertEquals(doctorAppointmentMock, service.findById(1L).get());
 	}
 	
 	@Test
