--- medicalcompany/src/main/java/com/personal/medical/repository/DentalAppointmentRepository.java
@@ -1,9 +1,0 @@
-package com.personal.medical.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.personal.medical.model.DentalAppointment;
-
-public interface DentalAppointmentRepository extends CrudRepository<DentalAppointment, Long> {
-
-}
