--- medicalcompany/src/main/java/com/personal/medical/repository/DoctorAppointmentRepository.java
@@ -1,9 +1,0 @@
-package com.personal.medical.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.personal.medical.model.DoctorAppointment;
-
-public interface DoctorAppointmentRepository extends CrudRepository<DoctorAppointment, Long> {
-
-}
