--- medicalcompany/src/main/java/com/personal/medical/services/DentalAppointmentService.java
@@ -1,8 +1,0 @@
-package com.personal.medical.services;
-
-import com.personal.medical.model.DentalAppointment;
-
-public interface DentalAppointmentService extends CrudService<DentalAppointment, Long> {
-	
-	
-}
