--- medicalcompany/src/main/java/com/personal/medical/services/DoctorAppointmentService.java
@@ -1,8 +1,0 @@
-package com.personal.medical.services;
-
-import com.personal.medical.model.DoctorAppointment;
-
-public interface DoctorAppointmentService extends CrudService<DoctorAppointment, Long> {
-	
-	
-}
